ds_1_array_ex_02 forces ndarray uses dtype=np.uint8

+++ ds_1_array/ex02/load_image.py
@@ -49,7 +49,7 @@ def ft_load(path: str) -> np.ndarray:
     """
     abspath = path_test(path)
     with Image.open(abspath) as im:
-        arr = np.array(im)
+        arr = np.array(im, dtype = np.uint8)
         print(f"The shape of image is: {arr.shape}")
         return arr
